Encapsulate playSound() in EffectsManager class
